fix: exit with an error exit code on error
The process always exits with a 0 exit code. In this case the tests
would never fail. There are obviously many more issues with this.

Now if there is an error thrown in the lua the process will exit with
the code of 1.

This will not exit 1 if one of the resource calls fails, an error will
need to be thrown on the lua side for this to exit with an error.

```lua
local ok = configz.directory "/some/dir"
assert(ok, "/some/dir was not created successfully")
```

It needs to be done this way because its OK for a `cmd` to exit with an
error. For situations you want to test to see if a programme is
installed before installing it.